requirements.txt
flask
flask-cors
pillow
tensorflow
scikit-learn
google-generativeai
fpdf
python-dotenv
gunicorn